requirements.txt
# Requirements for clackbot-qdb=api

flask
python-dotenv
psycopg2-binary